Created one-file library and verified
"nvc.a" gets compiled when IS_LIBRARY is set to 1. If it is moved into the "firmware" folder, then the contents of that file will be accessible without needing the source "*.cpp" file. Next step is to test this with multiple files that call on each other. It should be ensured that none of the intended files are excluded from the source. Hopefully it's all this simple lol or else we're kinda boned